+++ src/main/kotlin/app/revanced/patches/moneymanager/annotations/UnlockProCompatibility.kt
@@ -1,0 +1,9 @@
+package app.revanced.patches.moneymanager.annotations
+
+import app.revanced.patcher.annotation.Compatibility
+import app.revanced.patcher.annotation.Package
+
+@Compatibility([Package("com.ithebk.expensemanager")])
+@Target(AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.RUNTIME)
+internal annotation class UnlockProCompatibility