+++ src/main/kotlin/app/revanced/patches/sleepasandroid/annotations/UnlockPremiumCompatibility.kt
@@ -1,0 +1,9 @@
+package app.revanced.patches.sleepasandroid.annotations
+
+import app.revanced.patcher.annotation.Compatibility
+import app.revanced.patcher.annotation.Package
+
+@Compatibility([Package("com.urbandroid.sleep")])
+@Target(AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.RUNTIME)
+internal annotation class UnlockPremiumCompatibility