+++ src/main/kotlin/app/revanced/patches/twitch/ad/video/fingerprints/GetReadyToShowAdFingerprint.kt
@@ -1,9 +1,0 @@
-package app.revanced.patches.twitch.ad.video.fingerprints
-
-import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
-
-object GetReadyToShowAdFingerprint : MethodFingerprint(
-    customFingerprint = { method ->
-        method.definingClass.endsWith("/StreamDisplayAdsPresenter;") && method.name == "getReadyToShowAdOrAbort"
-    }
-)