+++ src/main/kotlin/app/revanced/patches/youtube/misc/fix/spoof/annotations/ClientSpoofCompatibility.kt
@@ -1,0 +1,14 @@
+package app.revanced.patches.youtube.misc.fix.spoof.annotations
+
+import app.revanced.patcher.annotation.Compatibility
+import app.revanced.patcher.annotation.Package
+
+@Compatibility(
+    [
+        Package("com.google.android.youtube"),
+        Package("com.vanced.android.youtube")
+    ]
+)
+@Target(AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.RUNTIME)
+internal annotation class ClientSpoofCompatibility